Trabaje sobre la logica del servicio y del controlador de turnos. Revisar si coincide con la regla de negocio y con lo que traeriamos del front. Hacer postman.

+++ src/main/java/ar/com/clinica/repository/ITurnoRepository.java
@@ -4,5 +4,5 @@ import ar.com.clinica.entity.Turno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ITurnoRepository extends JpaRepository<Turno, Integer> {
+public interface ITurnoRepository extends JpaRepository<Turno, Long> {
 }
